+++ lib/puma/events.rb
@@ -137,8 +137,24 @@ module Puma
       register(:on_booted, &block)
     end
 
+    def on_restart(&block)
+      register(:on_restart, &block)
+    end
+
+    def on_stopped(&block)
+      register(:on_stopped, &block)
+    end
+
     def fire_on_booted!
       fire(:on_booted)
+    end
+
+    def fire_on_restart!
+      fire(:on_restart)
+    end
+
+    def fire_on_stopped!
+      fire(:on_stopped)
     end
 
     DEFAULT = new(STDOUT, STDERR)