+++ lib/puma/launcher.rb
@@ -130,6 +130,7 @@ module Puma
 
     # Begin async shutdown of the server gracefully
     def stop
+      @events.fire_on_stopped!
       @status = :stop
       @runner.stop
     end
@@ -168,6 +169,7 @@ module Puma
 
       setup_signals
       set_process_title
+      integrate_with_systemd
       @runner.run
 
       case @status
@@ -242,6 +244,7 @@ module Puma
     end
 
     def restart!
+      @events.fire_on_restart!
       @config.run_hooks :on_restart, self, @events
 
       if Puma.jruby?
@@ -316,6 +319,30 @@ module Puma
       end
     end
 
+    #
+    # Puma's systemd integration allows Puma to inform systemd:
+    #  1. when it has successfully started
+    #  2. when it is starting shutdown
+    #  3. periodically for a liveness check with a watchdog thread
+    #
+
+    def integrate_with_systemd
+      return unless ENV["NOTIFY_SOCKET"]
+
+      begin
+        require 'puma/systemd'
+      rescue LoadError
+        log "Systemd integration failed. It looks like you're trying to use systemd notify but don't have sd_notify gem installed"
+        return
+      end
+
+      log "* Enabling systemd notification integration"
+
+      systemd = Systemd.new(@events)
+      systemd.hook_events
+      systemd.start_watchdog
+    end
+
     def spec_for_gem(gem_name)
       Bundler.rubygems.loaded_specs(gem_name)
     end
@@ -338,6 +365,7 @@ module Puma
     end
 
     def graceful_stop
+      @events.fire_on_stopped!
       @runner.stop_blocked
       log "=== puma shutdown: #{Time.now} ==="
       log "- Goodbye!"